requirements.txt
baidu_aip==4.16.13
Django==5.0.7
django-simple-captcha==0.6.0
playsound==1.3.0
Requests==2.32.3
chardet==5.2.0